Create classes for FDEM 1D
