+++ simpeg_drivers/electromagnetics/frequency_domain_1d/driver.py
@@ -7,3 +7,27 @@
 #  (see LICENSE file at the root of this source code package).                      '
 #                                                                                   '
 # '''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''
+
+
+from __future__ import annotations
+
+from simpeg_drivers.electromagnetics.base_1d_driver import Base1DDriver
+
+from .params import (
+    FDEM1DForwardOptions,
+    FDEM1DInversionOptions,
+)
+
+
+class FDEM1DForwardDriver(Base1DDriver):
+    """Frequency Domain 1D Electromagnetic forward driver."""
+
+    _params_class = FDEM1DForwardOptions
+    _validations = None
+
+
+class FDEM1DInversionDriver(Base1DDriver):
+    """Frequency Domain 1D Electromagnetic inversion driver."""
+
+    _params_class = FDEM1DInversionOptions
+    _validations = None
